Make findQuickPick return undefined when no QuickPick is found
Currently findQuickPick() returns a Promise<QuickPickItem|void> which is
inconvenient to use as you cannot use the return value directly without
casting. For instance an assertion or the ? operator only work on `null` or
`undefined` but not on `void`.

This commit changes the return type to be consistent with the Array.find()
function and returns now a Promise<QuickPickItem|undefined>, which removes the
above inconveniences.