Dont click tree items to expand them, use expand instead (#252)
* Remove trailing whitespace

* Add public function to expand TreeItems without clicking on them

This allows to expand items that have a command attached to them

* Use expand() in TreeItem.getChildItems() instead of click()

Clicking on the tree item only works for tree items that have no command
attached to them. For those that have a command bound to them, we have to use
expand(), as we would otherwise only trigger the command but not expand it.